+++ mRemoteNG/UI/Menu/msMain/ToolsMenu.cs
@@ -4,6 +4,7 @@ using System.Windows.Forms;
 using mRemoteNG.App;
 using mRemoteNG.Credential;
 using mRemoteNG.Resources.Language;
+using mRemoteNG.UI.Forms;
 
 namespace mRemoteNG.UI.Menu
 {
@@ -14,6 +15,7 @@ namespace mRemoteNG.UI.Menu
         private ToolStripMenuItem _mMenToolsExternalApps;
         private ToolStripMenuItem _mMenToolsPortScan;
         private ToolStripMenuItem _mMenToolsUvncsc;
+        private ToolStripMenuItem _mMenToolsMasterPassword;
 
         public Form MainForm { get; set; }
         public ICredentialRepositoryList CredentialProviderCatalog { get; set; }
@@ -29,6 +31,7 @@ namespace mRemoteNG.UI.Menu
             _mMenToolsUvncsc = new ToolStripMenuItem();
             _mMenToolsExternalApps = new ToolStripMenuItem();
             _mMenToolsPortScan = new ToolStripMenuItem();
+            _mMenToolsMasterPassword = new ToolStripMenuItem();
             // 
             // mMenTools
             // 
@@ -37,7 +40,9 @@ namespace mRemoteNG.UI.Menu
                 _mMenToolsSshTransfer,
                 _mMenToolsUvncsc,
                 _mMenToolsExternalApps,
-                _mMenToolsPortScan
+                _mMenToolsPortScan,
+                new ToolStripSeparator(),
+                _mMenToolsMasterPassword
             });
             Name = "mMenTools";
             Size = new System.Drawing.Size(48, 20);
@@ -74,6 +79,14 @@ namespace mRemoteNG.UI.Menu
             _mMenToolsPortScan.Size = new System.Drawing.Size(184, 22);
             _mMenToolsPortScan.Text = Language.PortScan;
             _mMenToolsPortScan.Click += mMenToolsPortScan_Click;
+            //
+            // mMenToolsMasterPassword
+            //
+            _mMenToolsMasterPassword.Image = Properties.Resources.Key_16x;
+            _mMenToolsMasterPassword.Name = "mMenToolsMasterPassword";
+            _mMenToolsMasterPassword.Size = new System.Drawing.Size(184, 22);
+            _mMenToolsMasterPassword.Text = Language.MasterPassword;
+            _mMenToolsMasterPassword.Click += mMenToolsMasterPassword_Click;
         }
 
         public void ApplyLanguage()
@@ -82,6 +95,7 @@ namespace mRemoteNG.UI.Menu
             _mMenToolsSshTransfer.Text = Language.SshFileTransfer;
             _mMenToolsExternalApps.Text = Language.ExternalTool;
             _mMenToolsPortScan.Text = Language.PortScan;
+            _mMenToolsMasterPassword.Text = Language.MasterPassword;
         }
 
         #region Tools
@@ -109,6 +123,12 @@ namespace mRemoteNG.UI.Menu
         private void mMenToolsOptions_Click(object sender, EventArgs e)
         {
             AppWindows.Show(WindowType.Options);
+        }
+
+        private void mMenToolsMasterPassword_Click(object sender, EventArgs e)
+        {
+            using MasterPasswordManager masterPasswordManager = new();
+            masterPasswordManager.ShowDialog(MainForm);
         }
 
         #endregion